refactor: added index.ts for server/helpers
