+++ sites/geohub/src/lib/server/helpers/fetchVectorInfo.ts
