+++ sites/geohub/src/routes/api/vector/statistics/+server.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from './$types'
 import { error } from '@sveltejs/kit'
-import { fetchVectorTileInfo } from '$lib/helper/fetchVectorInfo'
+import { fetchVectorTileInfo } from '$lib/server/helpers'
 
 export const GET: RequestHandler = async ({ url }) => {
   if (
